linters

+++ brew.py
@@ -1,5 +1,10 @@
-import os, subprocess, dotbot, sys
-from typing import Iterable, Mapping, Any
+import os
+import re
+import subprocess
+from collections.abc import Iterable, Mapping
+from typing import Any
+
+import dotbot
 
 Context = Mapping[str, Any]
 
@@ -48,10 +53,9 @@ class Brew(dotbot.Plugin):
         return self.directives[directive](data, context)
 
     def _tap(self, tap_list: Iterable[str], context: Context) -> bool:
-        if not self._brew_exist(context):
-            if not self._install_brew(context):
-                self._log.error("Failed to install brew")
-                return False
+        if not self._brew_exist(context) and not self._install_brew(context):
+            self._log.error("Failed to install brew")
+            return False
 
         result = True
 
@@ -65,14 +69,17 @@ class Brew(dotbot.Plugin):
         return result
 
     def _brew(self, packages_list: Iterable[str], context: Context):
-        if not self._brew_exist(context):
-            if not self._install_brew(context):
-                self._log.error("Failed to install brew")
-                return False
+        if not self._brew_exist(context) and not self._install_brew(context):
+            self._log.error("Failed to install brew")
+            return False
 
         result = True
 
         for pkg in packages_list:
+            if not self._valid_package(pkg):
+                self._log.warning(f"Invalid package name [{pkg}]")
+                result = False
+                continue
             self._log.info(f"Install {pkg}")
             success = self._invoke_shell_command(f"brew install {pkg}", context)
 
@@ -83,10 +90,9 @@ class Brew(dotbot.Plugin):
         return result
 
     def _brew_file(self, brew_files: Iterable[str], context: Context):
-        if not self._brew_exist(context):
-            if not self._install_brew(context):
-                self._log.error("Failed to install brew")
-                return False
+        if not self._brew_exist(context) and not self._install_brew(context):
+            self._log.error("Failed to install brew")
+            return False
 
         result = True
 
@@ -101,14 +107,18 @@ class Brew(dotbot.Plugin):
         return result
 
     def _cask(self, packages: Iterable[str], context: Context) -> bool:
-        if not self._brew_exist(context):
-            if not self._install_brew(context):
-                self._log.error("Failed to install brew")
-                return False
+        if not self._brew_exist(context) and not self._install_brew(context):
+            self._log.error("Failed to install brew")
+            return False
 
         result = True
 
         for pkg in packages:
+            if not self._valid_package(pkg):
+                self._log.warning(f"Invalid cask name [{pkg}]")
+                result = False
+                continue
+
             self._log.info(f"Install cask {pkg}")
             success = self._invoke_shell_command(
                 f"brew install --cask {pkg} || brew ls --cask --versions {pkg}",
@@ -121,14 +131,18 @@ class Brew(dotbot.Plugin):
         return result
 
     def _mas(self, apps: Iterable[str], context: Context):
-        if not self._mas_exist(context):
-            if not self._install_mas(context):
-                self._log.error("Failed to install mas")
-                return False
+        if not self._mas_exist(context) and not self._install_mas(context):
+            self._log.error("Failed to install mas")
+            return False
 
         result = True
 
         for app in apps:
+            if not self._valid_package(app):
+                self._log.warning(f"Invalid app name [{app}]")
+                result = False
+                continue
+
             self._log.info(f"Install app {app} for appStore")
             success = self._invoke_shell_command(f"mas install {app}", context)
 
@@ -146,10 +160,14 @@ class Brew(dotbot.Plugin):
 
     def _install_brew(self, context: Context) -> bool:
         link = "https://raw.githubusercontent.com/Homebrew/install/master/install.sh"
-        cmd = """hash brew || /bin/bash -c "$(curl -fsSL {0})";
-              brew update""".format(link)
+        cmd = f"""hash brew || /bin/bash -c "$(curl -fsSL {link})";
+              brew update"""
 
         return self._invoke_shell_command(cmd, context)
+
+    def _valid_package(self, package: str) -> bool:
+        result = re.match(r"^[a-zA-Z0-9/-]+$", package)
+        return result is not None
 
     def _install_mas(self, context: Context) -> bool:
         return self._brew(["mas"], context)
